pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nexo-infra"
version = "0.0.1"
description = "Infra package for Nexo"
authors = [
    { name = "Agra Bima Yuda", email = "agra@nexmedis.com" }
]
license = { text = "Proprietary"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types>=0.7.0",
    "anyio>=4.10.0",
    "bcrypt>=4.3.0",
    "black>=25.1.0",
    "cachetools>=5.5.2",
    "certifi>=2025.8.3",
    "cffi>=2.0.0",
    "cfgv>=3.4.0",
    "charset-normalizer>=3.4.3",
    "click>=8.3.0",
    "colorama>=0.4.6",
    "cryptography>=46.0.1",
    "distlib>=0.4.0",
    "fastapi>=0.116.2",
    "filelock>=3.19.1",
    "google-api-core>=2.25.1",
    "google-auth>=2.40.3",
    "google-cloud-appengine-logging>=1.6.2",
    "google-cloud-audit-log>=0.3.2",
    "google-cloud-core>=2.4.3",
    "google-cloud-logging>=3.12.1",
    "google-cloud-pubsub>=2.31.1",
    "googleapis-common-protos>=1.70.0",
    "greenlet>=3.2.4",
    "grpc-google-iam-v1>=0.14.2",
    "grpcio>=1.75.0",
    "grpcio-status>=1.75.0",
    "h11>=0.16.0",
    "httpcore>=1.0.9",
    "httpx>=0.28.1",
    "identify>=2.6.14",
    "idna>=3.10",
    "importlib_metadata>=8.7.0",
    "iniconfig>=2.1.0",
    "mypy_extensions>=1.1.0",
    "nexo-crypto>=0.0.2",
    "nexo-enums>=0.0.2",
    "nexo-logging>=0.0.3",
    "nexo-schemas>=0.0.1",
    "nexo-types>=0.0.2",
    "nexo-utils>=0.0.2",
    "nodeenv>=1.9.1",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-sdk>=1.37.0",
    "opentelemetry-semantic-conventions>=0.58b0",
    "packaging>=25.0",
    "pathspec>=0.12.1",
    "platformdirs>=4.4.0",
    "pluggy>=1.6.0",
    "pre_commit>=4.3.0",
    "proto-plus>=1.26.1",
    "protobuf>=6.32.1",
    "psutil>=7.1.0",
    "pyasn1>=0.6.1",
    "pyasn1_modules>=0.4.2",
    "pycparser>=2.23",
    "pycryptodome>=3.23.0",
    "pydantic>=2.12.2",
    "pydantic-settings>=2.10.1",
    "pydantic_core>=2.41.4",
    "Pygments>=2.19.2",
    "PyJWT>=2.10.1",
    "pytest>=8.4.2",
    "python-dotenv>=1.1.1",
    "pytokens>=0.1.10",
    "PyYAML>=6.0.2",
    "requests>=2.32.5",
    "rsa>=4.9.1",
    "sniffio>=1.3.1",
    "SQLAlchemy>=2.0.44",
    "starlette>=0.48.0",
    "typing-inspection>=0.4.2",
    "typing_extensions>=4.15.0",
    "ua-parser>=1.0.1",
    "ua-parser-builtins>=0.18.0.post1",
    "urllib3>=2.5.0",
    "user-agents>=2.2.0",
    "virtualenv>=20.34.0",
    "zipp>=3.23.0",
]

[tool.setuptools]
packages = [
    "nexo.infra",
    "nexo.infra.heartbeat",
    "nexo.infra.request",
    "nexo.infra.resource"
]

[tool.setuptools.package-data]
"nexo.infra" = ["*.json", "*.yaml"]
"nexo.infra.heartbeat" = ["*.json", "*.yaml"]
"nexo.infra.request" = ["*.json", "*.yaml"]
"nexo.infra.resource" = ["*.json", "*.yaml"]

[tool.setuptools.package-dir]
"nexo.infra" = "src"
"nexo.infra.heartbeat" = "src/heartbeat"
"nexo.infra.request" = "src/request"
"nexo.infra.resource" = "src/resource"
